+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Fix the default JavaScript timeout being too high (was 83 minutes, is now 5 seconds)
+
 ## [0.8.0] - 2019-02-01
 ### Changed
 - Rename the scrolling steps to use "scroll to middle" as the default. #23
@@ -68,6 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add "JsonApiSteps" trait for sending, receiving and asserting JSON data
 
+[Unreleased]: https://github.com/exozet/behat-utils/compare/0.8.0...HEAD
 [0.8.0]: https://github.com/exozet/behat-utils/compare/0.7.1...0.8.0
 [0.7.1]: https://github.com/exozet/behat-utils/compare/0.7.0...0.7.1
 [0.7.0]: https://github.com/exozet/behat-utils/compare/0.6.0...0.7.0